In-memory Sqlite database implemented
Database is created with Calculations table
Calculation results can be stored in database
Calculations table is logged each time data is saved - this is just to
demonstrate for this example that the data has been saved
DatabaseConnection is a singleton in order to reuse the database
connection.  Sqlite doesn't do multi-threading so database connection
pool or similar approaches are not relevant.